databse set input

+++ database/migrations/2022_11_02_061344_create_daftar_gaji_table.php
@@ -17,9 +17,9 @@ return new class extends Migration
             $table->increments('gaji_id');
             $table->integer('karyawan_id')->unsigned()->nullable();
             $table->foreign('karyawan_id')->references('karyawan_id')->on('karyawan')->cascadeOnDelete('cascade');
-            $table->bigInteger('Gaji_Pokok')->nullable();
-            $table->bigInteger('Pajak_Bpjs')->nullable();
-            $table->bigInteger('Jumlah_Gaji')->nullable();
+            $table->decimal('Gaji_Pokok')->nullable();
+            $table->decimal('Pajak_Bpjs')->nullable();
+            $table->decimal('Jumlah_Gaji')->nullable();
             $table->timestamps();
         });
     }
